Altternate method for drawing shadows
This adds an alternate method for drawing shadows that works for breakable boxes, but doesn’t look quite as nice for larger offsets.

With this commit the method for drawing shadow is automatically choosen based on the `breakable` option, but it might be preferrable to have a consistent look for all boxes and only use one method.